refactor: FSD Phase 1 - shared層の整備
- ui/ → shared/ui/ に移動（Modal, LoadingSpinner, ErrorDisplay）
- hooks/ → shared/lib/hooks/ に移動（useModal, useWindowSize）
- 各コンポーネントをディレクトリ化（index.ts追加）
- importパスを更新

FSD (Feature-Sliced Design) 移行の第一段階として
shared層を整備し、再利用可能なコンポーネントを集約
