+++ apps/frontend/src/shared/index.ts
@@ -1,0 +1,3 @@
+// shared - 共通リソース
+export * from './ui';
+export * from './lib';
