+++ apps/frontend/src/shared/lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useModal } from './useModal';
+export { useWindowSize } from './useWindowSize';
