+++ apps/frontend/src/shared/lib/hooks/useModal.ts
