+++ apps/frontend/src/shared/lib/hooks/useWindowSize.ts
