+++ apps/frontend/src/shared/ui/ErrorDisplay/index.ts
@@ -1,0 +1,2 @@
+export { ErrorDisplay, InlineError } from './ErrorDisplay';
+export type { ErrorDisplayProps } from './ErrorDisplay';
