+++ apps/frontend/src/shared/ui/LoadingSpinner/index.ts
@@ -1,0 +1,2 @@
+export { LoadingSpinner, InlineSpinner } from './LoadingSpinner';
+export type { LoadingSpinnerProps } from './LoadingSpinner';
