+++ apps/frontend/src/shared/ui/Modal/Modal.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react';
 import { createPortal } from 'react-dom';
-import { useModal } from '../hooks/useModal';
+import { useModal } from '../../lib/hooks/useModal';
 
 export interface ModalProps {
   isOpen: boolean;
