+++ apps/frontend/src/shared/ui/Modal/index.ts
@@ -1,0 +1,2 @@
+export { Modal } from './Modal';
+export type { ModalProps } from './Modal';
