+++ apps/frontend/src/shared/ui/index.ts
@@ -1,0 +1,9 @@
+// shared/ui - 汎用UIコンポーネント
+export { Modal } from './Modal';
+export type { ModalProps } from './Modal';
+
+export { LoadingSpinner, InlineSpinner } from './LoadingSpinner';
+export type { LoadingSpinnerProps } from './LoadingSpinner';
+
+export { ErrorDisplay, InlineError } from './ErrorDisplay';
+export type { ErrorDisplayProps } from './ErrorDisplay';
